fix: move azure specific code to azure auth package (#730)
